Volgende dingen veranderd: - Na laatste vraag kan er niet meer geklikt worden - Als de tijd op is dan komt er een popup
