+++ app/src/main/res/layout/activity_result.xml
@@ -85,6 +85,7 @@
                 android:textColor="#000"
                 android:padding="10dp"
                 android:textSize="20sp"
+                android:layout_marginEnd="10dp"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_weight="1"/>
@@ -94,6 +95,7 @@
                 android:textAllCaps="false"
                 android:textColor="#000"
                 android:padding="10dp"
+                android:layout_marginStart="10dp"
                 android:textSize="20sp"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
